requirements.txt
python-telegram-bot
stripe
python-decouple
flask